requirements.txt
# Database
asyncpg==0.29.0
alembic==1.13.1
SQLAlchemy==2.0.28

# API Framework
fastapi==0.110.0
uvicorn==0.27.1

# Testing
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==6.1.1
pytest-mock==3.14.0

# Utilities
python-jose[cryptography]==3.3.0  # For JWT handling
passlib[bcrypt]==1.7.4  # For password hashing
python-multipart==0.0.9  # For form data handling
pydantic==2.6.3
python-dotenv==1.0.1
pytz==2024.1  # For timezone handling

# AI/ML and Search (for agents that need them)
langchain>=0.1.0
langgraph>=0.0.15
langchain-anthropic>=0.1.0
langchain-community>=0.0.10
pgvector>=0.2.3
llama-parse>=0.1.0
langchain-core>=0.1.0
duckduckgo-search>=4.1.1
wikipedia>=1.4.0
supabase>=2.3.0
langsmith>=0.0.69
anthropic>=0.8.0

# Web scraping and HTTP
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Development tools
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
ruff>=0.1.0 